backend/src/spotter_eld_backend/requirements.txt
ipython
Django
mongoengine
redis
Pillow
phonenumbers
requests
mysqlclient
django-cors-headers
pytz
django-cors-headers
django-compressor
sentry-sdk
djangorestframework
djangorestframework-simplejwt